Added skeleton for code generation, changed license, ...
The code for the parser needs to be changed so that it preserves
information about what is a concatenation and what is an alternation,
it currently discards this information.

The license has changed, whilst the VM library will be LGPL, the
generator will not be.

References, comments and assertions need to be added throughout
the code.