Starting working on main test application.
The work on "rle.c" and "lzss.c" is nearly done, in terms of functionality,
there is a lot to do on other aspects and adding more helper functions.

"rewind" functionality has been added to "io.c".

Start on a new compressor function that pipes data from input, to RLE, then to
LZSS, then to output has begun, although it is not ready yet.