Added documentation, ideas for next step.
The next steps would be to generate C, but not for a virtual machine (this
would come later), it should be easier to generate a EBNF parser by directly
generating C statements to parse things. It is clear that this project is just
a first attempt and should be rewritten at a later date.